+++ objectified-ui/public/WHATS_NEW.md
@@ -30,6 +30,8 @@ We continue to improve the platform based on your feedback with improvements and
     - Adds TLS security options
     - Adds custom security schema options
   - Added ability to apply multiple security schemes to HTTP Operations (and/or)
+  - Added ability to specify different security scopes on a per-operation basis
+  - Configure required OAuth2/OpenID Connect scopes per operation (#421)
   - Dragging a property to a variable chip in the header will apply that property's primitive settings to the bound variable
 - OpenAPI Support:
   - Now at 100% support for class and properties definitions